WIP: Started migrating plugin to EmbeddedV2
