+++ android/src/main/java/co/pushe/plus/flutter/PusheBackgroundChannelHandler.java
@@ -1,0 +1,14 @@
+package co.pushe.plus.flutter;
+
+import android.support.annotation.NonNull;
+
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+
+public class PusheBackgroundChannelHandler implements MethodChannel.MethodCallHandler {
+
+    @Override
+    public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
+
+    }
+}
