+++ example/android/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
          additional functionality it is fine to subclass or reimplement
          FlutterApplication and put your custom class here. -->
     <application
-        android:name="io.flutter.app.FlutterApplication"
+        android:name="co.pushe.plus.flutter.PusheApplication"
         android:label="Pushe Flutter"
         android:icon="@mipmap/ic_launcher"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
@@ -31,6 +31,20 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+        <!-- Support Flutter embedding v1 activity -->
+        <activity
+            android:name=".EmbeddingV1Activity"
+            android:theme="@style/LaunchTheme"
+            android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale"
+            android:hardwareAccelerated="true"
+            android:windowSoftInputMode="adjustResize">
+        </activity>
+
+
+        <meta-data android:name="flutterEmbedding" android:value="2"/>
+
+        <!-- Pushe -->
         <meta-data
             android:name="pushe_token"
             android:value="OWdsNjQxeDV6cHJycjNxZ0Brb2xAMTU0MTQ0ODQ1NTY=" />
