+++ example/android/app/src/main/java/co/pushe/pushesampleflutter/MainActivity.java
@@ -1,15 +1,8 @@
 package co.pushe.pushesampleflutter;
 
-import android.os.Bundle;
 
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
 
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
 }
