+++ lib/pushe.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:ui';
 
 import 'package:flutter/services.dart';
 
@@ -11,6 +12,10 @@ enum IdType {
   AndroidId,
   GoogleAdvertisingId,
   CustomId
+}
+
+void _pusheSetupBackgroundChannel() {
+  
 }
 
 ///
@@ -26,7 +31,7 @@ class Pushe {
   static void Function(NotificationData, NotificationButtonData)
       _buttonClickCallback;
 
-  static const MethodChannel _channel = const MethodChannel('Pushe');
+  static const MethodChannel _channel = const MethodChannel('plus.pushe.co/pushe_flutter');
 
   /// Get the unique id of the devices generated by android id or ad id
   static Future<String> getAndroidId() async =>
@@ -224,6 +229,8 @@ class Pushe {
     _buttonClickCallback = onButtonClicked;
     _customContentCallback = onCustomContentReceived;
     _channel.setMethodCallHandler(_handleMethod);
+    final CallbackHandle backgroundSetupHandle = PluginUtilities.getCallbackHandle(_pusheSetupBackgroundChannel);
+
     if (!applicationOverridden) {
       _channel.invokeMethod("Pushe.initNotificationListenerManually");
     }
